# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Edubee Camp API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      security:
        - BearerAuth: []
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
            - super_admin
            - admin
            - camp_coordinator
            - education_agent
            - partner_institute
            - partner_hotel
            - partner_pickup
            - partner_tour
            - parent_client
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        lineId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        preferredLang:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        businessRegNo:
          type: string
          nullable: true
        countryOfOps:
          type: string
          nullable: true
        platformCommRate:
          type: string
          nullable: true
        status:
          type: string
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - role
        - fullName
        - status
        - createdAt
        - updatedAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        lineId:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        preferredLang:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        businessRegNo:
          type: string
          nullable: true
        countryOfOps:
          type: string
          nullable: true
        platformCommRate:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
      required:
        - email
        - password
        - role
        - fullName
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
        fullName:
          type: string
        phone:
          type: string
          nullable: true
        whatsapp:
          type: string
          nullable: true
        lineId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        preferredLang:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        businessRegNo:
          type: string
          nullable: true
        countryOfOps:
          type: string
          nullable: true
        platformCommRate:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
    UserListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...